Remove audio recorder component
