+++ src/SIL.XForge.Scripture/ClientApp/src/app/checking/checking/checking-answers/checking-answers.component.ts
@@ -37,7 +37,7 @@ import {
 import { QuestionDialogData } from '../../question-dialog/question-dialog.component';
 import { QuestionDialogService } from '../../question-dialog/question-dialog.service';
 import { TextAndAudioComponent } from '../../text-and-audio/text-and-audio.component';
-import { AudioAttachment } from '../checking-audio-recorder/checking-audio-recorder.component';
+import { AudioAttachment } from '../checking-audio-player/checking-audio-player.component';
 import { CheckingTextComponent } from '../checking-text/checking-text.component';
 import { CommentAction } from './checking-comments/checking-comments.component';
 import { CheckingQuestionComponent } from './checking-question/checking-question.component';
