--- src/SIL.XForge.Scripture/ClientApp/src/app/checking/checking/checking-audio-recorder/checking-audio-recorder.component.html
@@ -1,45 +1,0 @@
-<ng-container *transloco="let t; read: 'checking_audio_recorder'">
-  @if (!hasAudioAttachment) {
-    <div class="no-attachment">
-      @if (!isRecording) {
-        <button
-          mat-button
-          (click)="startRecording()"
-          class="record"
-          [matTooltip]="t('record_audio')"
-          matTooltipPosition="before"
-        >
-          <mat-icon>mic</mat-icon>
-        </button>
-      } @else {
-        <button
-          mat-button
-          (click)="stopRecording()"
-          class="stop"
-          [matTooltip]="t('stop_recording')"
-          matTooltipPosition="before"
-        >
-          <mat-icon>stop</mat-icon>
-        </button>
-      }
-    </div>
-  }
-  @if (hasAudioAttachment) {
-    <div class="has-attachment">
-      @if (!!audio.url) {
-        <app-single-button-audio-player [source]="audio.url" (click)="toggleAudio()">
-          <mat-icon>{{ audioPlayer?.playing ? "stop" : "play_arrow" }}</mat-icon>
-        </app-single-button-audio-player>
-      }
-      <button
-        mat-button
-        (click)="resetRecording()"
-        [matTooltip]="t('delete')"
-        class="remove-audio-file"
-        matTooltipPosition="before"
-      >
-        <mat-icon>close</mat-icon>
-      </button>
-    </div>
-  }
-</ng-container>
